+++ Changelog.md
@@ -11,7 +11,7 @@
 1. Added GUI overlay plugins. Users can now write a Gazebo + QT plugin that displays widgets over the render window.
   * [Pull request #1181](https://bitbucket.org/osrf/gazebo/pull-request/1181)
 
-1. Change behavior of Joint::SetVelocity
+1. Change behavior of Joint::SetVelocity, add Joint::SetVelocityLimit(unsigned int, double)
   * [Pull request]()
 
 ## Gazebo 4.0